requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
cobble==0.1.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
Django==5.0.7
django-cors-headers==4.4.0
django-rest-authtoken==2.1.4
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.1.0rc8
filetype==1.2.0
future==1.0.0
gunicorn==23.0.0
idna==3.8
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
mammoth==1.8.0
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
pycparser==2.22
pycryptodome==3.20.0
python-docx==1.1.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
thumbnail==1.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
zipfile36==0.1.3
